chatbot/backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-openai
langchain-community
openai
sentence-transformers
huggingface-hub
pypdf2
python-multipart
python-dotenv
pydantic
pydantic-settings
httpx
aiofiles
python-jose[cryptography]
passlib[bcrypt]
pandas
numpy
tiktoken
faiss-cpu
langsmith
